add action handler; move stuff to common
